core: Split resolve traits into core (#317)
The proxy's service discovery is aroudn the proxy::Resolve traits. In
order to support moving discovery logic into a dedicated subcrate, these
traits should be promoted into the core types library.

The `ListenAndSpawn` and `ServeConnection` APIs have been moved into a
`listen` submodule.
