+++ src/app/proxy.rs
@@ -1,4 +1,5 @@
-use crate::core::{drain, ListenAndSpawn, ServeConnection};
+use crate::core::listen::{ListenAndSpawn, ServeConnection};
+use crate::drain;
 use futures::{self, Future, Poll};
 use tracing::error;
