+++ src/proxy/server.rs
@@ -1,5 +1,5 @@
 use super::Accept;
-use crate::core::ServeConnection;
+use crate::core::listen::ServeConnection;
 use crate::proxy::http::{
     glue::{HttpBody, HyperServerSvc},
     upgrade,
